% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{check_noaa_present_arrow}
\alias{check_noaa_present_arrow}
\title{Check if NOAA forecasts have been downloaded and processed}
\usage{
check_noaa_present_arrow(
  lake_directory,
  configure_run_file = "configure_run.yml",
  config_set_name = "default"
)
}
\arguments{
\item{lake_directory}{four-letter code for site}

\item{configure_run_file}{name of simulation}

\item{config_set_name}{FLARE configuration object (needed for s3 buckets and endpoit)}
}
\description{
Check if NOAA forecasts have been downloaded and processed
}
