% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flow_files.R
\name{create_flow_files}
\alias{create_flow_files}
\title{Generating a list of flow files in the flare_tempdir}
\usage{
create_flow_files(
  flow_forecast_dir = NULL,
  flow_historical_dir = NULL,
  flow_type = "inflow",
  variables = c("time", "FLOW", "TEMP", "SALT"),
  out_dir,
  start_datetime,
  end_datetime = NA,
  forecast_start_datetime = NA,
  forecast_horizon = 0,
  site_id,
  use_s3 = FALSE,
  bucket = NULL,
  endpoint = NULL,
  local_directory = NULL,
  use_ler_vars = FALSE,
  config = config,
  out_dir_fn = NULL
)
}
\arguments{
\item{flow_forecast_dir}{location of the forecast files}

\item{flow_historical_dir}{location of the historical files}

\item{flow_type}{inflow or outflow}

\item{variables}{what variables are included in the flow file}

\item{out_dir}{the directory in which to put the flow files (e.g. flare_tempdir)}

\item{start_datetime}{start of simulation}

\item{end_datetime}{end of simulation}

\item{forecast_start_datetime}{start of the forecast period (break between historical + future periods)}

\item{forecast_horizon}{horizon}

\item{site_id}{site code}

\item{use_s3}{logical}

\item{bucket}{s3 storage location}

\item{endpoint}{s3 storage location}

\item{local_directory}{local storage location}

\item{use_ler_vars}{T/F}

\item{config}{list of FLARE configurations, used for S3 access when \code{use_s3 = TRUE}}

\item{out_dir_fn}{Optional function of the ensemble positional index (integer,
1-based) returning the directory for that member's flow file. When non-NULL
each file is written to \code{out_dir_fn(i)}; when NULL files are written
to \code{out_dir}. Default \code{NULL}.}
}
\value{
matrix of flow_file_names
}
\description{
Generating a list of flow files in the flare_tempdir
}
