% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_glm_inflow_outflow_files.R
\name{create_glm_inflow_outflow_files}
\alias{create_glm_inflow_outflow_files}
\title{Generating inflow and output files in the GLM format}
\usage{
create_glm_inflow_outflow_files(
  inflow_file_dir,
  inflow_obs,
  working_directory,
  config,
  state_names
)
}
\arguments{
\item{inflow_file_dir}{string; full directory path that contains forecasted inflow and outflow files}

\item{inflow_obs}{string; full path to cleaned inflow observation in the specified format}

\item{working_directory}{string; full directory where FLARE executes}

\item{state_names}{vector; vector of state names that will be included in the inflow files}
}
\value{
list with two vectors. One vector is the matrix of inflow_file_names and the other is the matrix of outflow_file_names
}
\description{
Generating inflow and output files in the GLM format
}
\details{
Processes historical inflow data from inflow_obs and from files in the inflow_file_dir into the GLM format
}
\examples{
\dontrun{
inflow_outflow_files <- create_glm_inflow_outflow_files(inflow_file_dir = inflow_forecast_path, inflow_obs = cleaned_inflow_file, working_directory = config$file_path$execute_directory, config, state_names = NULL)
}
}
