% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inflow_outflow_file.R
\name{create_inflow_outflow_file}
\alias{create_inflow_outflow_file}
\title{Create inflow and outflow files in GLM format}
\usage{
create_inflow_outflow_file(
  full_time_local,
  working_directory,
  input_file_tz = "EST5EDT",
  start_forecast_step,
  inflow_file1,
  inflow_file2,
  chemistry_file,
  local_tzone,
  met_file_names,
  forecast_days,
  inflow_process_uncertainty,
  future_inflow_flow_coeff,
  future_inflow_flow_error,
  future_inflow_temp_coeff,
  future_inflow_temp_error,
  states_config,
  include_wq,
  use_future_inflow,
  doc_scalar
)
}
\arguments{
\item{doc_scalar}{}
}
\value{
None
}
\description{
Create inflow and outflow files in GLM format
}
\author{
Quinn Thomas
}
