% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inflow_outflow_files_arrow.R
\name{create_inflow_outflow_files_arrow}
\alias{create_inflow_outflow_files_arrow}
\title{Generating inflow and outflow files in the GLM format using arrow}
\usage{
create_inflow_outflow_files_arrow(config, config_set_name)
}
\arguments{
\item{config}{configuration file}
}
\value{
list with two vectors. One vector is the matrix of inflow_file_names and the other is the matrix of outflow_file_names
}
\description{
Generating inflow and outflow files in the GLM format using arrow
}
\details{
Processes historical model data and future model files into the GLM format
}
