% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_matrix.R
\name{create_obs_matrix}
\alias{create_obs_matrix}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
create_obs_matrix(
  cleaned_observations_file_long,
  config,
  start_datetime_local,
  end_datetime_local,
  local_tzone,
  modeled_depths
)
}
\arguments{
\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
