% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_met_input.R
\name{create_obs_met_input}
\alias{create_obs_met_input}
\title{Create met file using observed meteorology}
\usage{
create_obs_met_input(
  fname,
  outfile,
  full_time_local,
  local_tzone,
  working_directory,
  hist_days,
  missing_met_data_threshold
)
}
\arguments{
\item{fname}{}

\item{missing_met_data_threshold}{}
}
\description{
Create met file using observed meteorology
}
\author{
Quinn Thomas
}
