% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sss_input_output.R
\name{create_sss_input_output}
\alias{create_sss_input_output}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
create_sss_input_output(
  x,
  i,
  m,
  full_time_local,
  working_directory,
  wq_start,
  management_input,
  hist_days,
  forecast_sss_on,
  sss_depth,
  use_specified_sss,
  states_config,
  include_wq
)
}
\arguments{
\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
