% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{delete_restart}
\alias{delete_restart}
\title{Delete restart file on s3 bucket}
\usage{
delete_restart(site, sim_name)
}
\arguments{
\item{sim_name}{}
}
\value{

}
\description{
Delete restart file on s3 bucket
}
