% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{delete_sim}
\alias{delete_sim}
\title{Title}
\usage{
delete_sim(site_id, sim_name)
}
\arguments{
\item{site_id}{four letter code for site}

\item{sim_name}{name of simulation}
}
\value{

}
\description{
Title
}
