% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{download_s3_objects}
\alias{download_s3_objects}
\title{Download file from s3 bucket}
\usage{
download_s3_objects(lake_directory, bucket, prefix)
}
\arguments{
\item{prefix}{}
}
\value{

}
\description{
Download file from s3 bucket
}
