% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flare_io.R
\name{flare_delete_file}
\alias{flare_delete_file}
\title{Delete a single remote object.}
\usage{
flare_delete_file(remote_file, server_name = "", remote_folder = "", config)
}
\arguments{
\item{remote_file}{Remote object filename.}

\item{server_name}{Name of the DataStore (key under \code{config$s3}).}

\item{remote_folder}{Remote folder/prefix.}

\item{config}{FLAREr config list.}
}
\value{
\code{TRUE} on success (invisibly).
}
\description{
Delete a single remote object.
}
