% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flare_io.R
\name{flare_put_file}
\alias{flare_put_file}
\title{Upload a single local file as an object.}
\usage{
flare_put_file(
  local_file,
  remote_file,
  server_name = "",
  local_folder = ".",
  remote_folder = ".",
  config
)
}
\arguments{
\item{local_file}{Local destination filename.}

\item{remote_file}{Remote object filename.}

\item{server_name}{Name of the DataStore (key under \code{config$s3}).}

\item{local_folder}{Local folder for \code{local_file}.}

\item{remote_folder}{Remote folder/prefix for \code{remote_file}.}

\item{config}{FLAREr config list.}
}
\description{
Upload a single local file as an object.
}
