% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_forecast_score.R
\name{generate_forecast_score}
\alias{generate_forecast_score}
\title{Score a forecast using score4cast package}
\usage{
generate_forecast_score(
  targets_file,
  forecast_file,
  output_directory,
  variable_types = "state"
)
}
\arguments{
\item{targets_file}{observation file}

\item{forecast_file}{forecast file}
}
\description{
Score a forecast using score4cast package
}
