% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_glm_met_files.R
\name{generate_glm_met_files}
\alias{generate_glm_met_files}
\title{Convert NOAA forecasts to GLM format}
\usage{
generate_glm_met_files(
  obs_met_file = NULL,
  out_dir,
  forecast_dir = NULL,
  local_tzone,
  start_datetime_UTC,
  end_datetime_UTC,
  forecast_start_datetime_UTC
)
}
\arguments{
\item{forecast_start_datetime_UTC}{}
}
\value{

}
\description{
Convert NOAA forecasts to GLM format
}
