% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_glm_met_files_arrow.R
\name{generate_glm_met_files_arrow}
\alias{generate_glm_met_files_arrow}
\title{Convert historical meteorology and NOAA forecasts to GLM format via arrow connection}
\usage{
generate_glm_met_files_arrow(
  obs_met_file = NULL,
  out_dir,
  forecast_dir = NULL,
  config
)
}
\arguments{
\item{obs_met_file}{string; full path to netcdf that is observed historical meteorology}

\item{out_dir}{string; full path to directory where the converted files will be saved}

\item{forecast_dir}{string; full path to arrow file system with the NOAA forecast netcdf files}
}
\value{
list; vector of full path for the converted files and boolean flag if issues with historical meteorology files
}
\description{
Convert historical meteorology and NOAA forecasts to GLM format via arrow connection
}
\details{
Function combines historical meteorology and NOAA forecasts to create meteorology input files in the GLM format.  A file is generated for each ensemble member.
}
\examples{
\dontrun{
met_out <- FLAREr::generate_glm_met_files(obs_met_file = observed_met_file, out_dir = config$file_path$execute_directory, forecast_dir = config$file_path$noaa_directory, config)
}
}
\author{
Quinn Thomas
}
