% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_conditions_ler.R
\name{generate_initial_conditions_ler}
\alias{generate_initial_conditions_ler}
\title{Generate initial conditions from observations and default values}
\usage{
generate_initial_conditions_ler(
  states_config,
  obs_config,
  pars_config = NULL,
  obs,
  config,
  model
)
}
\arguments{
\item{config}{}
}
\value{

}
\description{
Generate initial conditions from observations and default values
}
