% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ler_met_files.R
\name{generate_ler_met_files}
\alias{generate_ler_met_files}
\title{Convert NOAA forecasts to GLM format}
\usage{
generate_ler_met_files(
  obs_met_file = NULL,
  out_dir,
  forecast_dir = NULL,
  local_tzone,
  start_datetime_local,
  end_datetime_local,
  forecast_start_datetime_local
)
}
\arguments{
\item{local_tzone}{}
}
\value{

}
\description{
Convert NOAA forecasts to GLM format
}
