% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_met_files.R
\name{generate_met_files}
\alias{generate_met_files}
\title{Convert NOAA forecasts to GLM format}
\usage{
generate_met_files(obs_met_file = NULL, out_dir, forecast_dir = NULL, config)
}
\arguments{
\item{use_ler}{}
}
\value{

}
\description{
Convert NOAA forecasts to GLM format
}
