% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_states_to_obs_mapping.R
\name{generate_states_to_obs_mapping}
\alias{generate_states_to_obs_mapping}
\title{Convert columns the map states to observations to a single vector}
\usage{
generate_states_to_obs_mapping(states_config, obs_config)
}
\arguments{
\item{states_config}{}
}
\value{

}
\description{
Convert columns the map states to observations to a single vector
}
