% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_driver_forecast}
\alias{get_driver_forecast}
\title{Download driver forecasts from s3 bucket}
\usage{
get_driver_forecast(lake_directory, forecast_path)
}
\arguments{
\item{forecast_path}{}
}
\value{

}
\description{
Download driver forecasts from s3 bucket
}
