% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_driver_forecast_path}
\alias{get_driver_forecast_path}
\title{Get file path for driver forecasts}
\usage{
get_driver_forecast_path(config, forecast_model)
}
\arguments{
\item{forecast_model}{}
}
\value{

}
\description{
Get file path for driver forecasts
}
