% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_edi_file}
\alias{get_edi_file}
\title{Download file from EDI data portal}
\usage{
get_edi_file(edi_https, file, lake_directory)
}
\arguments{
\item{lake_directory}{}
}
\value{

}
\description{
Download file from EDI data portal
}
