% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_nml_functions.R
\name{get_glm_nc_var_all_wq}
\alias{get_glm_nc_var_all_wq}
\title{Add together two numbers.}
\usage{
get_glm_nc_var_all_wq(
  ncFile,
  working_dir,
  z_out,
  vars_depth,
  vars_no_depth,
  diagnostic_vars
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}
