% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_restart_file}
\alias{get_restart_file}
\title{Download restart file from s3 bucket}
\usage{
get_restart_file(config, lake_directory)
}
\arguments{
\item{lake_directory}{}
}
\value{

}
\description{
Download restart file from s3 bucket
}
