% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_run_config}
\alias{get_run_config}
\title{Get run configuration from s3 bucket}
\usage{
get_run_config(configure_run_file, lake_directory, config, clean_start)
}
\arguments{
\item{clean_start}{}
}
\value{

}
\description{
Get run configuration from s3 bucket
}
