% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_stacked_noaa}
\alias{get_stacked_noaa}
\title{Download stacked NOAA data from s3 bucket}
\usage{
get_stacked_noaa(lake_directory, config, averaged = TRUE)
}
\arguments{
\item{averaged}{}
}
\value{

}
\description{
Download stacked NOAA data from s3 bucket
}
