% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_targets}
\alias{get_targets}
\title{Download target data from s3}
\usage{
get_targets(lake_directory, config)
}
\arguments{
\item{config}{}
}
\value{

}
\description{
Download target data from s3
}
