% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_general.R
\name{plotting_general}
\alias{plotting_general}
\title{Plotting FLARE output}
\usage{
plotting_general(file_name, qaqc_location)
}
\arguments{
\item{file_name}{}

\item{qaqc_location}{}
}
\value{
None
}
\description{
Plotting FLARE output
}
\author{
Quinn Thomas
}
