% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_general_2.R
\name{plotting_general_2}
\alias{plotting_general_2}
\title{Generate diagnostic plot of FLARE output with observations}
\usage{
plotting_general_2(
  file_name,
  target_file,
  ncore = 1,
  plot_profile = FALSE,
  obs_csv = FALSE
)
}
\arguments{
\item{file_name}{string; full path to output file produced by write_forecast_netcdf()}

\item{ncore}{integer; number of computer cores for parallel processing}

\item{plot_profile}{boolean; include profile plots in PDF}

\item{obs_csv}{boolean; generate csv with the observed and predicted values for model states}

\item{qaqc_data_directory}{string; full path to processed long-format observation file}
}
\value{
None
}
\description{
Generate diagnostic plot of FLARE output with observations
}
\details{
Function combines the netcdf output with the long-format observation file to produce a set of plots for each state variable, calibrated parameter, and diagnostic variable
}
\examples{
\dontrun{
plotting_general_2(file_name = saved_file, target_file = target_file, ncore = config$model_settings$ncore, plot_profile = TRUE, obs_csv)
}
}
\author{
Quinn Thomas
}
