% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmtools.R
\name{print.nml}
\alias{print.nml}
\title{Print NML}
\usage{
\method{print}{nml}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"nml"}.}

\item{...}{Ignored.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print NML
}
