% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_downscale_GEFS.R
\name{process_downscale_GEFS}
\alias{process_downscale_GEFS}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
process_downscale_GEFS(
  folder,
  noaa_location,
  input_met_file,
  working_directory,
  n_ds_members,
  n_met_members,
  file_name,
  local_tzone,
  FIT_PARAMETERS,
  DOWNSCALE_MET,
  met_downscale_uncertainty,
  compare_output_to_obs,
  VarInfo,
  replaceObsNames,
  downscaling_coeff,
  full_time_local,
  first_obs_date,
  last_obs_date,
  input_met_file_tz,
  weather_uncertainty,
  obs_met_outfile,
  lake_latitude,
  lake_longitude
)
}
\arguments{
\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
