% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_parameters.R
\name{propose_parameters}
\alias{propose_parameters}
\title{Propose new parameters for use in data assimilation and forecasting}
\usage{
propose_parameters(
  i,
  m,
  pars,
  pars_config,
  npars,
  par_fit_method,
  da_method,
  hist_days,
  include_uncertainty
)
}
\arguments{
\item{i}{time step index}

\item{m}{ensemble member number}

\item{pars}{matrix of parameters}

\item{pars_config}{parameter configuration data frame}

\item{npars}{number of parameters}

\item{par_fit_method}{method for parameter fitting}

\item{da_method}{data assimilation method}

\item{hist_days}{number of simulation days before forecasting}

\item{include_uncertainty}{include parameter uncertainty in forecasts}
}
\value{
vector of new parameter values for the ensemble member
}
\description{
Propose new parameters for use in data assimilation and forecasting
}
