% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{put_forecast}
\alias{put_forecast}
\title{Upload forecast file and metadata to s3 bucket}
\usage{
put_forecast(saved_file, eml_file_name, config)
}
\arguments{
\item{config}{}
}
\value{

}
\description{
Upload forecast file and metadata to s3 bucket
}
