% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{put_restart_file}
\alias{put_restart_file}
\title{Upload restart netcdf file to s3 bucket}
\usage{
put_restart_file(saved_file, config)
}
\arguments{
\item{saved_file}{full path of saved FLARE netcdf}

\item{config}{flare configuration object}
}
\description{
Upload restart netcdf file to s3 bucket
}
\keyword{internal}
