% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{put_score}
\alias{put_score}
\title{Upload score to s3 bucket}
\usage{
put_score(saved_file, config)
}
\arguments{
\item{config}{flare configuration object}

\item{score_file}{full path of saved FLARE netcdf}
}
\description{
Upload score to s3 bucket
}
