% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{put_targets}
\alias{put_targets}
\title{Save target files to s3 bucket}
\usage{
put_targets(
  site_id,
  cleaned_insitu_file = NA,
  cleaned_met_file = NA,
  cleaned_inflow_file = NA,
  use_s3
)
}
\arguments{
\item{cleaned_inflow_file}{}
}
\value{

}
\description{
Save target files to s3 bucket
}
