% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmtools.R
\name{read_nml}
\alias{read_nml}
\title{Read a GLM/ELCOM namelist file into an R list}
\usage{
read_nml(nml_file = "template")
}
\arguments{
\item{nml_file}{Path to a \code{.nml} file.}
}
\value{
A named list of class \code{"nml"} with one element per namelist block.
}
\description{
Read a GLM/ELCOM namelist file into an R list
}
