% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_enkf.R
\name{run_enkf}
\alias{run_enkf}
\title{Title}
\usage{
run_enkf(
  x_matrix,
  h,
  pars_corr,
  zt,
  psi_t,
  z_index,
  states_depth_start,
  states_height_start,
  model_internal_heights_start,
  lake_depth_start,
  log_particle_weights_start,
  snow_ice_thickness_start,
  avg_surf_temp_start,
  mixer_count_start,
  mixing_vars_start,
  diagnostics_start,
  pars_config,
  config,
  depth_index,
  depth_obs,
  depth_sd,
  par_fit_method
)
}
\arguments{
\item{par_fit_method}{}
}
\description{
Title
}
