% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model_glm_aed.R
\name{run_model}
\alias{run_model}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
run_model(
  i,
  m,
  mixing_vars_start,
  curr_start,
  curr_stop,
  par_names,
  curr_pars,
  working_directory,
  par_nml,
  num_phytos,
  glm_depths_start,
  surface_height_start,
  simulate_SSS,
  x_start,
  full_time_local,
  wq_start,
  wq_end,
  management_input,
  hist_days,
  forecast_sss_on,
  sss_depth,
  use_specified_sss,
  modeled_depths,
  ndepths_modeled,
  curr_met_file,
  inflow_file_names,
  inflow_outflow_index,
  outflow_file_names,
  glm_output_vars,
  diagnostics_names,
  machine,
  npars,
  num_wq_vars,
  snow_ice_thickness_start,
  avg_surf_temp_start,
  nstates,
  states_config,
  include_wq,
  specified_sss_inflow_file
)
}
\arguments{
\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
