% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_ler.R
\name{run_models_ler}
\alias{run_models_ler}
\title{Export output settings for each model}
\usage{
run_models_ler(model, folder, verbose)
}
\arguments{
\item{model}{vector; model to export configuration file.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{folder}

\item{config_file}{name of the master LakeEnsemblR config file}
}
\description{
Exports settings related to output (time step, format),
for each model
}
\examples{

}
\keyword{methods}
