% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_particile_filter.R
\name{run_particle_filter}
\alias{run_particle_filter}
\title{Title}
\usage{
run_particle_filter(
  x_matrix,
  h,
  pars_corr,
  zt,
  psi,
  z_index,
  states_depth_start,
  states_height_start,
  model_internal_heights_start,
  lake_depth_start,
  log_particle_weights_start,
  snow_ice_thickness_start,
  avg_surf_temp_start,
  mixer_count_start,
  mixing_vars_start,
  diagnostics_start,
  pars_config,
  config,
  depth_index,
  depth_obs,
  depth_sd,
  par_fit_method,
  vertical_obs
)
}
\arguments{
\item{vertical_obs}{}
}
\description{
Title
}
