% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{set_configuration}
\alias{set_configuration}
\title{Set and create directories in the configuration file}
\usage{
set_configuration(configure_run_file, lake_directory, clean_start = FALSE)
}
\arguments{
\item{clean_start}{}
}
\value{

}
\description{
Set and create directories in the configuration file
}
