% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model_glm_aed.R
\name{set_up_model}
\alias{set_up_model}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
set_up_model(
  executable_location,
  working_directory,
  base_GLM_nml,
  num_wq_vars,
  base_AED_nml,
  base_AED_phyto_pars_nml,
  base_AED_zoop_pars_nml,
  ndepths_modeled,
  modeled_depths,
  the_sals_init,
  machine,
  include_wq
)
}
\arguments{
\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
\keyword{internal}
