% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_simulation.R
\name{set_up_simulation}
\alias{set_up_simulation}
\title{Set and create directories in the configuration file}
\usage{
set_up_simulation(
  configure_run_file = "configure_run.yml",
  lake_directory,
  clean_start = FALSE,
  config_set_name = "default",
  sim_name = NA
)
}
\arguments{
\item{configure_run_file}{name of run configuration file (do not include full path)}

\item{lake_directory}{full path to repository directory}

\item{clean_start}{logical: TRUE = reset run configuration with the file in the configuration directory within repository}

\item{config_set_name}{name of configuration set}

\item{sim_name}{name of simulation}
}
\value{
list of configuration values (Invisibly); side effect of creating necessary subdirectories in \code{lake_directory}
}
\description{
Set and create directories in the configuration file
}
\examples{
dir <- normalizePath(tempdir(),  winslash = "/")
lake_directory <- file.path(dir, "extdata")
# Copy files to temporarly directory
dir.create(dir,showWarnings = FALSE)
file.copy(system.file("extdata", package = "FLAREr"),
          tempdir(),
          recursive = TRUE)

set_up_simulation(configure_run_file = "configure_run.yml",
                  lake_directory = lake_directory,
                   clean_start = FALSE,
                   config_set_name = "default",
                   sim_name = NA)



}
