% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_geom.R
\name{solar_geom}
\alias{solar_geom}
\title{solar_geom}
\usage{
solar_geom(doy, lon, lat)
}
\arguments{
\item{doy}{time as day of year. Integers indicated midnight.}

\item{lon}{longitude}

\item{lat}{latitude}
}
\description{
calculates potential top-of-atmosphere shortwave radiation as a function of day of year and location
}
\author{
Mike Dietze
}
