% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_nml_functions.R
\name{update_phyto}
\alias{update_phyto}
\title{Add together two numbers.}
\usage{
update_phyto(p_initial, nml_name = "aed2_phyto_pars.nml")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}
