% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{update_run_config2}
\alias{update_run_config2}
\title{Update run configuration}
\usage{
update_run_config2(
  lake_directory,
  configure_run_file = "configure_run.yml",
  restart_file,
  start_datetime,
  end_datetime,
  forecast_start_datetime,
  forecast_horizon,
  sim_name,
  site_id,
  configure_flare,
  configure_obs,
  use_s3,
  bucket,
  endpoint,
  use_https = TRUE
)
}
\arguments{
\item{use_https}{}
}
\description{
Update run configuration
}
