% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{update_run_config_neon}
\alias{update_run_config_neon}
\title{Update run configuration and upload to s3 bucket}
\usage{
update_run_config_neon(
  config,
  lake_directory,
  configure_run_file = "configure_run.yml"
)
}
\arguments{
\item{config}{flare configuration object}

\item{lake_directory}{full path to repository directory}

\item{configure_run_file}{name of run configuration file (do not include full path)}
}
\description{
Update run configuration and upload to s3 bucket
}
