% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_nml_functions.R
\name{update_var}
\alias{update_var}
\title{Add together two numbers.}
\usage{
update_var(var_value, var_name, working_directory, nml)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}
