% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_forecast.R
\name{write_forecast}
\alias{write_forecast}
\title{Generate parquet output file}
\usage{
write_forecast(
  da_forecast_output,
  use_s3 = FALSE,
  bucket = NULL,
  endpoint = NULL,
  local_directory = NULL
)
}
\arguments{
\item{da_forecast_output}{list; object that is returned by run_da_forecast()}

\item{use_s3}{Boolen; use s3 storage for saving scores}

\item{bucket}{S3 bucket}

\item{endpoint}{S3 endpoint}

\item{local_directory}{local directory of scores if not using s3}
}
\value{
None
}
\description{
Generate parquet output file
}
\details{
Function generates a parquet file from the object that is returned by run_da_forecast()
}
\author{
Quinn Thomas
}
\keyword{internal}
