% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_forecast_arrow.R
\name{write_forecast_arrow}
\alias{write_forecast_arrow}
\title{Generate parquet output file}
\usage{
write_forecast_arrow(
  da_forecast_output,
  use_s3 = FALSE,
  bucket = NULL,
  endpoint = NULL,
  local_directory = NULL
)
}
\arguments{
\item{da_forecast_output}{list; object that is returned by run_da_forecast()}

\item{forecast_output_directory}{string; full path of directory where the csv file will be written}

\item{use_short_filename}{use shortened file name; this results in less information in the file name and potentially overwriting existing files}
}
\value{
None
}
\description{
Generate parquet output file
}
\details{
Function generates a netcdf file from the object that is returned by run_da_forecast()
}
\author{
Quinn Thomas
}
