% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_forecast_netcdf.R
\name{write_forecast_netcdf}
\alias{write_forecast_netcdf}
\title{Generate netcdf output file}
\usage{
write_forecast_netcdf(da_forecast_output, forecast_output_directory)
}
\arguments{
\item{da_forecast_output}{list; object that is returned by run_da_forecast()}

\item{forecast_output_directory}{string; full path of directory where the netcdf file will be written}
}
\value{
None
}
\description{
Generate netcdf output file
}
\details{
Function generates a netcdf file from the object that is returned by run_da_forecast()
}
\examples{
\dontrun{
write_forecast_netcdf(da_forecast_output = da_forecast_output, forecast_output_directory = config$file_path$forecast_output_directory)
}

}
\author{
Quinn Thomas
}
