% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_forecast_parquet.R
\name{write_forecast_parquet}
\alias{write_forecast_parquet}
\title{Generate parquet output file}
\usage{
write_forecast_parquet(
  da_forecast_output,
  forecast_output_directory,
  use_short_filename = TRUE
)
}
\arguments{
\item{da_forecast_output}{list; object that is returned by run_da_forecast()}

\item{forecast_output_directory}{string; full path of directory where the csv file will be written}

\item{use_short_filename}{use shortened file name; this results in less information in the file name and potentially overwriting existing files}
}
\value{
None
}
\description{
Generate parquet output file
}
\details{
Function generates a netcdf file from the object that is returned by run_da_forecast()
}
\examples{
\dontrun{
write_forecast_csv(da_forecast_output = da_forecast_output, forecast_output_directory = config$file_path$forecast_output_directory, use_short_filename = TRUE)
}

}
\author{
Quinn Thomas
}
