% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmtools.R
\name{write_nml}
\alias{write_nml}
\title{Write an nml list to a namelist file}
\usage{
write_nml(glm_nml, file)
}
\arguments{
\item{glm_nml}{A named list of class \code{"nml"} (from \code{read_nml}).}

\item{file}{Path of the output \code{.nml} file.}
}
\value{
Invisibly, \code{NULL}.
}
\description{
Write an nml list to a namelist file
}
