% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.R
\name{write_restart}
\alias{write_restart}
\title{Generate netcdf restart file}
\usage{
write_restart(
  da_forecast_output,
  forecast_output_directory,
  use_short_filename = TRUE
)
}
\arguments{
\item{da_forecast_output}{list; object that is returned by run_da_forecast()}

\item{forecast_output_directory}{string; full path of directory where the netcdf file will be written}

\item{use_short_filename}{use shortened file name; this results in less informatoin in the file name and potentially overwriting existing files}
}
\value{
None
}
\description{
Generate netcdf restart file
}
\details{
Function generates a netcdf file from the object that is returned by run_da_forecast()
}
\author{
Quinn Thomas
}
\keyword{internal}
