% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{check_noaa_present}
\alias{check_noaa_present}
\title{Check if NOAA forecasts have been downloaded and processed}
\usage{
check_noaa_present(
  lake_directory,
  configure_run_file = "configure_run.yml",
  config_set_name = "default"
)
}
\arguments{
\item{lake_directory}{four-letter code for site}

\item{configure_run_file}{name of simulation}

\item{config_set_name}{FLARE configuration object (needed for s3 buckets and endpoit)}
}
\value{
logical
}
\description{
Check if NOAA forecasts have been downloaded and processed
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

dir <- normalizePath(tempdir(),  winslash = "/")
lake_directory <- file.path(dir, "extdata")
# Copy files to temporarly directory
dir.create(dir,showWarnings = FALSE)
file.copy(system.file("extdata", package = "FLAREr"),
          tempdir(),
          recursive = TRUE)

check_noaa_present(lake_directory,
                   configure_run_file = "configure_run.yml",
                   config_set_name = "default")
\dontshow{\}) # examplesIf}
}
