% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{check_noaa_present}
\alias{check_noaa_present}
\title{Check if NOAA forecast is on s3 bucket}
\usage{
check_noaa_present(
  lake_directory,
  configure_run_file = "configure_run.yml",
  config_set_name = "default"
)
}
\arguments{
\item{lake_directory}{full path to repository directory}

\item{configuration_file}{file name (no path) of run configuration file}
}
\description{
Check if NOAA forecast is on s3 bucket
}
