% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_forecast_observations.R
\name{combine_forecast_observations}
\alias{combine_forecast_observations}
\title{Title}
\usage{
combine_forecast_observations(
  file_name,
  qaqc_location,
  extra_historical_days,
  ncore = 1
)
}
\arguments{
\item{extra_historical_days}{}
}
\value{

}
\description{
Title
}
