% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flare_metadata.R
\name{create_flare_metadata}
\alias{create_flare_metadata}
\title{Generate metadata (EML) using the Ecological Forecasting Initiative Standards}
\usage{
create_flare_metadata(file_name, da_forecast_output)
}
\arguments{
\item{da_forecast_output}{}
}
\value{
None
}
\description{
Generate metadata (EML) using the Ecological Forecasting Initiative Standards
}
\author{
Quinn Thomas
}
