% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inflow_outflow_file.R
\name{create_inflow_outflow_file}
\alias{create_inflow_outflow_file}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
create_inflow_outflow_file(
  full_time_local,
  working_directory,
  input_file_tz = "EST5EDT",
  start_forecast_step,
  inflow_file1,
  inflow_file2,
  outflow_file1,
  chemistry_file,
  local_tzone,
  met_file_names,
  forecast_days,
  inflow_process_uncertainty,
  future_inflow_flow_coeff,
  future_inflow_flow_error,
  future_inflow_temp_coeff,
  future_inflow_temp_error,
  states_config,
  include_wq,
  use_future_inflow,
  doc_scalar
)
}
\arguments{
\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
