% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inflow_outflow_files.R
\name{create_inflow_outflow_files}
\alias{create_inflow_outflow_files}
\title{Generating inflow and outflow files in the GLM format using arrow}
\usage{
create_inflow_outflow_files(config, config_set_name, lake_directory)
}
\arguments{
\item{config}{configuration file}

\item{config_set_name}{specific name of configuration within the configuration directory}

\item{lake_directory}{directory for FLARE application}
}
\value{
list with two vectors. One vector is the matrix of inflow_file_names and the other is the matrix of outflow_file_names
}
\description{
Generating inflow and outflow files in the GLM format using arrow
}
\details{
Processes historical model data and future model files into the GLM format
}
\keyword{internal}
