% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inflow_outflow_files_arrow.R,
%   R/create_inflow_outflow_files_arrow_old.R
\name{create_inflow_outflow_files_arrow}
\alias{create_inflow_outflow_files_arrow}
\title{Generating inflow and outflow files in the GLM format using arrow}
\usage{
create_inflow_outflow_files_arrow(
  inflow_forecast_dir = NULL,
  inflow_obs,
  variables = c("time", "FLOW", "TEMP", "SALT"),
  out_dir,
  start_datetime,
  end_datetime = NA,
  forecast_start_datetime = NA,
  forecast_horizon = 0,
  site_id,
  use_s3 = FALSE,
  bucket = NULL,
  endpoint = NULL,
  local_directory = NULL,
  use_forecast = TRUE,
  use_ler_vars = FALSE
)

create_inflow_outflow_files_arrow(
  inflow_forecast_dir = NULL,
  inflow_obs,
  variables = c("time", "FLOW", "TEMP", "SALT"),
  out_dir,
  start_datetime,
  end_datetime = NA,
  forecast_start_datetime = NA,
  forecast_horizon = 0,
  site_id,
  use_s3 = FALSE,
  bucket = NULL,
  endpoint = NULL,
  local_directory = NULL,
  use_forecast = TRUE,
  use_ler_vars = FALSE
)
}
\arguments{
\item{inflow_forecast_dir}{string; full directory path that contains forecasted inflow and outflow files}

\item{inflow_obs}{string; full path to cleaned inflow observation in the specified format}

\item{config}{configuration file}

\item{working_directory}{string; full directory where FLARE executes}

\item{state_names}{vector; vector of state names that will be included in the inflow files}
}
\value{
list with two vectors. One vector is the matrix of inflow_file_names and the other is the matrix of outflow_file_names

list with two vectors. One vector is the matrix of inflow_file_names and the other is the matrix of outflow_file_names
}
\description{
Generating inflow and outflow files in the GLM format using arrow

Generating inflow and output files in the GLM format using arrow
}
\details{
Processes historical model data and future model files into the GLM format

Processes historical inflow data from inflow_obs and from files in the inflow_file_dir into the GLM format
}
