% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ler_inflow_outflow_files.R
\name{create_ler_inflow_outflow_files}
\alias{create_ler_inflow_outflow_files}
\title{Title}
\usage{
create_ler_inflow_outflow_files(
  inflow_file_dir,
  inflow_obs,
  working_directory,
  start_datetime_local,
  end_datetime_local,
  forecast_start_datetime_local,
  use_future_inflow,
  state_names
)
}
\arguments{
\item{state_names}{}
}
\value{

}
\description{
Title
}
