% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_met_files.R
\name{create_met_files}
\alias{create_met_files}
\title{Convert historical meteorology and NOAA forecasts to GLM format}
\usage{
create_met_files(config, lake_directory)
}
\arguments{
\item{config}{list of FLARE configurations}

\item{lake_directory}{directory of lake configurations}
}
\value{
list; vector of full path for the converted files and boolean flag if issues with historical meteorology files
}
\description{
Convert historical meteorology and NOAA forecasts to GLM format
}
\details{
Function combines historical meteorology and NOAA forecasts to create meteorology input files in the GLM format.  A file is generated for each ensemble member.
}
\author{
Quinn Thomas
}
