% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_met_files_openmet.R
\name{create_met_files_openmet}
\alias{create_met_files_openmet}
\title{Generate GLM ready met files from open meteo}
\usage{
create_met_files_openmet(
  out_dir,
  start_datetime,
  end_datetime = NA,
  forecast_start_datetime = NA,
  forecast_horizon = 0,
  latitude,
  longitude,
  site_id,
  openmeteo_api = "ensemble",
  model = NULL,
  use_archive = FALSE,
  bucket = NULL,
  endpoint = NULL
)
}
\arguments{
\item{out_dir}{path to where the GLM formatted met file will be written}

\item{start_datetime}{start date of weather data}

\item{end_datetime}{end date of weather data}

\item{forecast_start_datetime}{start datetime of forecast}

\item{forecast_horizon}{number of days in the future}

\item{latitude}{latitude}

\item{longitude}{longitude (west is negative)}

\item{site_id}{site code}

\item{openmeteo_api}{type of weather data or forecast (ensemble_forecast, season, climate, historical)}

\item{model}{forecast model}

\item{use_archive}{Boolen (default = FALSE); TRUE = use forecasts stored on s3 bucket, FALSE = use open-meteo download directly}

\item{bucket}{s3 bucket for archive}

\item{endpoint}{s3 endpoint for archive}
}
\value{
list of meteorology file names
}
\description{
Generate GLM ready met files from open meteo
}
\details{
Generate GLM ready met files from open meteo
}
\keyword{internal}
