% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_matrix.R
\name{create_obs_matrix}
\alias{create_obs_matrix}
\title{Create matrix of observations in the format required by run_da_forecast}
\usage{
create_obs_matrix(cleaned_observations_file_long, obs_config, config)
}
\arguments{
\item{cleaned_observations_file_long}{file name (with full path)}
}
\value{
matrix
}
\description{
Create matrix of observations in the format required by run_da_forecast
}
\author{
Quinn Thomas
}
