% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_matrix.R
\name{create_obs_matrix}
\alias{create_obs_matrix}
\title{Create matrix of observations in the format required by run_ENKF}
\usage{
create_obs_matrix(
  cleaned_observations_file_long,
  config,
  start_datetime_local,
  end_datetime_local,
  local_tzone,
  modeled_depths
)
}
\arguments{
\item{cleaned_observations_file_long}{}

\item{modeled_depths}{}
}
\description{
Create matrix of observations in the format required by run_ENKF
}
\author{
Quinn Thomas
}
