% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_matrix.R
\name{create_obs_matrix}
\alias{create_obs_matrix}
\title{Create matrix of observations in the format required by run_da_forecast}
\usage{
create_obs_matrix(cleaned_observations_file_long, obs_config, config)
}
\arguments{
\item{cleaned_observations_file_long}{string; file name (with full path) of the long-format observation file}

\item{obs_config}{list; observations configuration list}

\item{config}{list; flare configuration list}
}
\value{
matrix that is based to generate_initial_conditions() and run_da_forecast()
}
\description{
Create matrix of observations in the format required by run_da_forecast
}
\details{
Creates a matrix of observations that maps the modeled states to the observed states. The function uses information from columns the obs_config file.
}
\author{
Quinn Thomas
}
\keyword{internal}
