% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_non_vertical.R
\name{create_obs_non_vertical}
\alias{create_obs_non_vertical}
\title{Process observation for variables that don't have depth}
\usage{
create_obs_non_vertical(
  cleaned_observations_file_long,
  obs_config,
  start_datetime,
  end_datetime,
  forecast_start_datetime,
  forecast_horizon,
  secchi_sd = 0.5,
  depth_sd = 0.1
)
}
\arguments{
\item{obs_config}{description}

\item{depth_sd}{}
}
\description{
Process observation for variables that don't have depth
}
