% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_non_vertical.R
\name{create_obs_non_vertical}
\alias{create_obs_non_vertical}
\title{Process observation for variables that don't have depth}
\usage{
create_obs_non_vertical(
  cleaned_observations_file_long,
  obs_config,
  start_datetime,
  end_datetime,
  forecast_start_datetime,
  forecast_horizon
)
}
\arguments{
\item{cleaned_observations_file_long}{string; file name (with full path) of the long-format observation file}

\item{obs_config}{string; file name (with full path) of the long-format observation file}

\item{start_datetime}{first datetime of the simulation}

\item{end_datetime}{last datetime of the simulation}

\item{forecast_start_datetime}{datetime that a forecast starts}

\item{forecast_horizon}{number of days forecasted}
}
\value{
list of secchi and depth observations for use in FLAREs
}
\description{
Process observation for variables that don't have depth
}
\details{
Process observation for variables that don't have depth
}
\keyword{internal}
