% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{delete_restart}
\alias{delete_restart}
\title{Delete restart file on s3 bucket}
\usage{
delete_restart(site_id, sim_name, bucket = "restart", endpoint)
}
\arguments{
\item{site_id}{four letter code for site}

\item{sim_name}{name of simulation}

\item{bucket}{s3 bucket}

\item{endpoint}{S3 endpoint}
}
\description{
Delete restart file on s3 bucket
}
\keyword{internal}
