% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{delete_sim}
\alias{delete_sim}
\title{Title}
\usage{
delete_sim(site_id, sim_name, config)
}
\arguments{
\item{site_id}{four letter code for site}

\item{sim_name}{name of simulation}

\item{config}{FLARE configuration object (needed for s3 buckets and endpoit)}
}
\description{
Title
}
