% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{download_s3_objects}
\alias{download_s3_objects}
\title{Download file from s3 bucket}
\usage{
download_s3_objects(lake_directory, bucket, prefix, region, base_url)
}
\arguments{
\item{lake_directory}{full path to repository directory}

\item{bucket}{name of s3 bucket}

\item{prefix}{relative path directory within bucket}

\item{region}{S3 region}

\item{base_url}{S3 endpoint}
}
\description{
Download file from s3 bucket
}
