% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_forecast_score.R
\name{generate_forecast_score}
\alias{generate_forecast_score}
\title{Generate forecast score from csv target and forecast files}
\usage{
generate_forecast_score(targets_file, forecast_file, output_directory)
}
\arguments{
\item{output_directory}{}
}
\value{

}
\description{
Generate forecast score from csv target and forecast files
}
