% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_forecast_score_arrow.R
\name{generate_forecast_score_arrow}
\alias{generate_forecast_score_arrow}
\title{Score a forecast using score4cast package and arrow}
\usage{
generate_forecast_score_arrow(
  targets_file,
  forecast_df,
  use_s3 = FALSE,
  bucket = NULL,
  endpoint = NULL,
  local_directory = NULL,
  variable_types = "state"
)
}
\arguments{
\item{targets_file}{observation file}

\item{forecast_df}{forecast file}
}
\description{
Score a forecast using score4cast package and arrow
}
