% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_conditions.R
\name{generate_initial_conditions}
\alias{generate_initial_conditions}
\title{Generate initial conditions for FLARE}
\usage{
generate_initial_conditions(
  states_config,
  obs_config,
  pars_config = NULL,
  obs,
  config,
  obs_non_vertical
)
}
\arguments{
\item{states_config}{list; list of state configurations}

\item{obs_config}{list; list of observation configurations}

\item{pars_config}{list; list of parameter configurations  (Default = NULL)}

\item{obs}{array; array of the observations. Required dimensions are \verb{[nobs, time, depth]}}

\item{config}{list; list of configurations}

\item{obs_non_vertical}{list; observations of variables that don't have vertical dimension.}
}
\value{
list; list contains the initial conditions objects required by run_da_forecast()
}
\description{
Generate initial conditions for FLARE
}
\details{
Function to generate initial conditions from either default values in the states_config, observations (if available), or a previous run using the output as a restart file.
}
\author{
Quinn Thomas
}
\keyword{internal}
