% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_conditions.R
\name{generate_initial_conditions}
\alias{generate_initial_conditions}
\title{Generate initial conditions from observations and default values}
\usage{
generate_initial_conditions(
  states_config,
  obs_config,
  pars_config = NULL,
  obs,
  config,
  restart_file = NA,
  historical_met_error = FALSE
)
}
\arguments{
\item{historical_met_error}{}
}
\value{

}
\description{
Generate initial conditions from observations and default values
}
