% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_LER_met_files.R
\name{generate_ler_met_files}
\alias{generate_ler_met_files}
\title{Convert NOAA forecasts to LER format}
\usage{
generate_ler_met_files(
  obs_met_file = NULL,
  out_dir,
  forecast_dir = NULL,
  config
)
}
\arguments{
\item{config}{}
}
\value{

}
\description{
Convert NOAA forecasts to LER format
}
