% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_met_files_arrow.R
\name{generate_met_files_arrow}
\alias{generate_met_files_arrow}
\title{Convert historical meteorology and NOAA forecasts to GLM format via arrow connection}
\usage{
generate_met_files_arrow(
  obs_met_file = NULL,
  out_dir,
  start_datetime,
  end_datetime = NA,
  forecast_start_datetime = NA,
  forecast_horizon = 0,
  site_id,
  use_s3 = FALSE,
  bucket = NULL,
  endpoint = NULL,
  local_directory = NULL,
  use_forecast = TRUE,
  use_ler_vars = FALSE,
  use_siteid_s3 = FALSE
)
}
\arguments{
\item{obs_met_file}{string; full path to netcdf that is observed historical meteorology}

\item{out_dir}{string; full path to directory where the converted files will be saved}
}
\value{
list; vector of full path for the converted files and boolean flag if issues with historical meteorology files
}
\description{
Convert historical meteorology and NOAA forecasts to GLM format via arrow connection
}
\details{
Function combines historical meteorology and NOAA forecasts to create meteorology input files in the GLM format.  A file is generated for each ensemble member.
}
\author{
Quinn Thomas
}
