% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oxygen_management.R
\name{generate_oxygen_management}
\alias{generate_oxygen_management}
\title{Generate management list required to simulate oxygen inputs}
\usage{
generate_oxygen_management(config)
}
\arguments{
\item{end_datetime}{}
}
\value{

}
\description{
Generate management list required to simulate oxygen inputs
}
