% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_restart_initial_conditions.R
\name{generate_restart_initial_conditions}
\alias{generate_restart_initial_conditions}
\title{Generate initial conditiosn from saved output file}
\usage{
generate_restart_initial_conditions(
  restart_file,
  state_names,
  par_names = NULL
)
}
\arguments{
\item{restart_file}{}

\item{par_names}{}
}
\description{
Generate initial conditiosn from saved output file
}
\author{
Quinn Thomas
}
