% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_states_to_obs_mapping.R
\name{generate_states_to_obs_mapping}
\alias{generate_states_to_obs_mapping}
\title{Map the states to the observations}
\usage{
generate_states_to_obs_mapping(states_config, obs_config)
}
\arguments{
\item{states_config}{list; list of state configurations}

\item{obs_config}{list; list of observation configurations}
}
\value{
updated states_config list
}
\description{
Map the states to the observations
}
\details{
Convert columns the map states to observations to a single vector.  This is necessary because not all states have observations, multiple states can contribute to the same observation, and one observation can be associated with multiple states. Function must be assigned to the states_config object
}
\author{
Quinn Thomas
}
\keyword{internal}
