% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_driver_forecast}
\alias{get_driver_forecast}
\title{Download driver forecasts from s3 bucket}
\usage{
get_driver_forecast(lake_directory, forecast_path)
}
\arguments{
\item{lake_directory}{full path to repository directory}

\item{forecast_path}{relative path of the driver forecast (relative to driver directory or bucket)}
}
\value{

}
\description{
Download driver forecasts from s3 bucket
}
