% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_git_repo}
\alias{get_git_repo}
\title{Get data from Github repository}
\usage{
get_git_repo(lake_directory, directory, git_repo)
}
\arguments{
\item{lake_directory}{full path to repository directory}

\item{directory}{the branch name on github}

\item{git_repo}{https of the github repository}
}
\description{
Get data from Github repository
}
