% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_restart_file}
\alias{get_restart_file}
\title{Download restart file from s3 bucket}
\usage{
get_restart_file(config, lake_directory)
}
\arguments{
\item{config}{flare configuration object}

\item{lake_directory}{full path to repository directory}
}
\description{
Download restart file from s3 bucket
}
