% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{get_run_config}
\alias{get_run_config}
\title{Get run configuration from s3 bucket}
\usage{
get_run_config(
  configure_run_file = "configure_run.yml",
  lake_directory,
  config,
  clean_start = FALSE,
  config_set_name = "default",
  sim_name = NA
)
}
\arguments{
\item{configure_run_file}{file name (no path) of run configuration file}

\item{lake_directory}{full path to repository directory}

\item{config}{flare configuration object}

\item{clean_start}{logical; reset the configuration run to the base file in the configuration directory}
}
\description{
Get run configuration from s3 bucket
}
