% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{initialize_obs_processing}
\alias{initialize_obs_processing}
\title{Set and create directories for observation configuration}
\usage{
initialize_obs_processing(
  lake_directory,
  observation_yml = NA,
  config_set_name = "default"
)
}
\arguments{
\item{lake_directory}{full path to repository directory}
}
\description{
Set and create directories for observation configuration
}
