% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_model_error.R
\name{initiate_model_error}
\alias{initiate_model_error}
\title{Initiate model error from configuration files}
\usage{
initiate_model_error(config, states_config)
}
\arguments{
\item{config}{list from reading in configuration yaml}

\item{states_config}{list from reading in states configuration csv}
}
\value{

}
\description{
Initiate model error from configuration files
}
