% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_general.R
\name{plotting_general}
\alias{plotting_general}
\title{Generate diagnostic plot of FLARE output with observations}
\usage{
plotting_general(forecast_df, targets_df, file_name)
}
\arguments{
\item{forecast_df}{dataframe; data frame containing the forecast predictions}

\item{targets_df}{dataframe; data frame containing all observations (targets)}

\item{file_name}{string; full path to file containing plotting output}
}
\value{
None
}
\description{
Generate diagnostic plot of FLARE output with observations
}
\details{
Function combines the netcdf output with the long-format observation file to produce a set of plots for each state variable, calibrated parameter, and diagnostic variable
}
\author{
Austin Delany
}
