% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_general.R
\name{plotting_general}
\alias{plotting_general}
\title{Generate diagnostic plot of FLARE output with observations}
\usage{
plotting_general(
  file_name,
  qaqc_data_directory,
  ncore = 1,
  plot_profile = TRUE,
  obs_csv = TRUE
)
}
\arguments{
\item{file_name}{string; full path to output file produced by write_forecast_netcdf()}

\item{qaqc_data_directory}{string; full path to processed long-format observation file}

\item{ncore}{integer; number of computer cores for parallel processing}

\item{plot_profile}{boolean; include profile plots in PDF}

\item{obs_csv}{boolean; generate csv with the observed and predicted values for model states}
}
\value{
None
}
\description{
Generate diagnostic plot of FLARE output with observations
}
\details{
Function combines the netcdf output with the long-format observation file to produce a set of plots for each state variable, calibrated parameter, and diagnostic variable
}
\author{
Quinn Thomas
}
