% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_downscale_GEFS.R
\name{process_downscale_GEFS}
\alias{process_downscale_GEFS}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
process_downscale_GEFS(
  noaa_location,
  input_met_file,
  working_directory,
  n_ds_members,
  n_met_members,
  file_name,
  local_tzone,
  FIT_PARAMETERS,
  DOWNSCALE_MET,
  met_downscale_uncertainty,
  compare_output_to_obs,
  VarInfo,
  replaceObsNames,
  downscaling_coeff,
  full_time_local,
  first_obs_date,
  last_obs_date,
  input_met_file_tz,
  weather_uncertainty,
  obs_met_outfile,
  lake_latitude,
  lake_longitude
)
}
\arguments{
\item{noaa_location}{}

\item{lake_longitude}{}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas

Laura Puckett
}
