% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{put_forecast_csv}
\alias{put_forecast_csv}
\title{Upload forecast csv to s3 bucket}
\usage{
put_forecast_csv(saved_file, config)
}
\arguments{
\item{config}{flare configuration object}

\item{score_file}{full path of saved FLARE netcdf}
}
\description{
Upload forecast csv to s3 bucket
}
