% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{put_targets}
\alias{put_targets}
\title{Save target files to s3 bucket}
\usage{
put_targets(
  site_id,
  cleaned_insitu_file = NA,
  cleaned_met_file = NA,
  cleaned_inflow_file = NA,
  use_s3 = FALSE,
  config
)
}
\arguments{
\item{site_id}{four letter code for the site}

\item{cleaned_insitu_file}{full path of the cleaned insitu file}

\item{cleaned_met_file}{full path of the cleaned met file}

\item{cleaned_inflow_file}{full path of the cleaned inflow file}

\item{use_s3}{logical; TRUE = use s3}
}
\description{
Save target files to s3 bucket
}
