% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_EnKF.R
\name{run_EnKF}
\alias{run_EnKF}
\title{Run Ensemble Kalman Filter}
\usage{
run_EnKF(
  states_init,
  pars_init,
  obs,
  obs_sd,
  model_sd,
  working_directory,
  met_file_names,
  inflow_file_names,
  outflow_file_names,
  sim_start_datetime,
  sim_end_datetime,
  forecast_start_datetime = NA,
  config,
  pars_config = NULL,
  states_config,
  obs_config,
  aux_states_init,
  management = NULL
)
}
\arguments{
\item{states_init, }{array of the initial states.  Required dimensions are \link{states, depths, ensemble}}

\item{pars_init, }{array of the initial states.  Required dimensions are \link{pars, depths, ensemble}}

\item{obs, }{array of the observaitons. Required dimensions are \link{nobs, time, depth}}

\item{model_sd, }{vector of standard deviations describing the model error for each state}

\item{working_directory, }{directory model executes}

\item{met_file_names, }{vector of meterology file names}

\item{inflow_file_names, }{vector of inflow file names}

\item{outflow_file_names, }{vector of outflow file names}

\item{sim_start_datetime, }{datetime of beginning of the simulation}

\item{sim_end_datetime, }{datetime of the end of the simulation}

\item{forecast_start_datetime, }{datetime when simulate is a forecast}

\item{config, }{list of configurations}

\item{pars_config, }{list of parameter configurations}

\item{states_config, }{list of state configurations}

\item{obs_config, }{list of observation configurations}

\item{aux_states_init, }{list of initial conditions for auxillary states}

\item{management.}{list of management inputs and configuration}
}
\value{
None
}
\description{
Run Ensemble Kalman Filter
}
\author{
Quinn Thomas
}
