% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_EnKF.R
\name{run_EnKF}
\alias{run_EnKF}
\title{Download and Downscale NOAA GEFS for a single site}
\usage{
run_EnKF(
  x_init,
  obs,
  psi,
  process_sd,
  working_directory,
  met_file_names,
  inflow_file_names,
  outflow_file_names,
  sim_start_datetime,
  sim_end_datetime,
  forecast_start_datetime = NA,
  management_input,
  wq_start,
  wq_end,
  config,
  aux_states_init,
  code_folder
)
}
\arguments{
\item{site_index, }{index of site_list, lat_list, lon_list to be downloaded}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{site_list, }{vector of site codes, used in directory and file name generation}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{model_name, }{directory name for the 6-hr forecast, this will be used in directory and file name generation}

\item{model_name_ds, }{directory name for the 1-hr forecast, this will be used in directory and file name generation}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Download and Downscale NOAA GEFS for a single site
}
\author{
Quinn Thomas
}
