% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_ler.R
\name{run_models_ler}
\alias{run_models_ler}
\title{Run the lake model from LakeEnsemblR}
\usage{
run_models_ler(
  model,
  folder,
  verbose,
  restart,
  member,
  the_temps,
  model_depths
)
}
\arguments{
\item{model}{vector; model to export configuration file.
Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")}

\item{folder}{folder}

\item{verbose}{Boolean; Should model output be shown in the console. Defaults to FALSE}
}
\description{
Run the lake model from LakeEnsemblR
}
\examples{

}
\keyword{methods}
