% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{set_configuration}
\alias{set_configuration}
\title{Set and create directories in the configuration file}
\usage{
set_configuration(
  configure_run_file = "configure_run.yml",
  lake_directory,
  clean_start = FALSE,
  config_set_name = "default",
  sim_name = NA
)
}
\arguments{
\item{configure_run_file}{name of run configuration file (do not include full path)}

\item{lake_directory}{full path to repository directory}

\item{clean_start}{logical: TRUE = reset run configuration with the file in the configuration directory within repository}
}
\description{
Set and create directories in the configuration file
}
