% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{update_run_config}
\alias{update_run_config}
\title{Update run configuration and upload to s3 bucket}
\usage{
update_run_config(
  config,
  lake_directory,
  configure_run_file,
  saved_file,
  new_horizon,
  day_advance = 1
)
}
\arguments{
\item{day_advance}{}
}
\value{

}
\description{
Update run configuration and upload to s3 bucket
}
