% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{update_run_config}
\alias{update_run_config}
\title{Update run configuration and upload to s3 bucket}
\usage{
update_run_config(
  config,
  lake_directory,
  configure_run_file = "configure_run.yml",
  saved_file = NA,
  new_horizon = NA,
  day_advance = NA,
  new_start_datetime = TRUE
)
}
\arguments{
\item{config}{flare configuration object}

\item{lake_directory}{full path to repository directory}

\item{configure_run_file}{name of run configuration file (do not include full path)}

\item{saved_file}{full path of saved FLARE netcdf}

\item{new_horizon}{horizon (in days) to update the run configuration with}

\item{day_advance}{number of days between forecast forecast generation (defaults to 1)}
}
\description{
Update run configuration and upload to s3 bucket
}
