% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_forecast_netcdf.R
\name{write_forecast_netcdf}
\alias{write_forecast_netcdf}
\title{Write netcdf output file from object created by run_ENKF}
\usage{
write_forecast_netcdf(enkf_output, forecast_location)
}
\arguments{
\item{enkf_output}{}

\item{forecast_location}{}
}
\value{
None
}
\description{
Write netcdf output file from object created by run_ENKF
}
\author{
Quinn Thomas
}
